+++ src/useSQLite.ts
@@ -142,7 +142,30 @@ export interface SQLiteHook extends AvailableResult {
      * @since 2.0.1
      */
     checkConnectionsConsistency(): Promise<Result>;
-
+    /**
+     * Check if secure secret has been stored
+     * @returns Promise<Result>
+     * @since 2.0.2
+     */
+    isSecretStored(): Promise<Result>; 
+    /**
+     * Set an encrypted secret to secure storage
+     * To run only once
+     * Will migrate from GlobalSQLite secret when required
+     * @param passphrase 
+     * @returns Promise<void>
+     * @since 2.0.2
+     */
+    setEncryptionSecret(passphrase: string): Promise<void>;   
+    /**
+     * Change encrypted secret from secure storage
+     * Not to use to migrate from GlobalSQLite secret (run setEncryptionSecret)
+     * @param passphrase 
+     * @param oldpassphrase 
+     * @returns Promise<void>
+     * @since 2.0.2
+     */
+    changeEncryptionSecret(passphrase: string, oldpassphrase: string): Promise<void>;  
 
 }
 
@@ -499,6 +522,47 @@ export const useSQLite = ({
         }
 
     }, [mSQLite]);
+    const isSecretStored = useCallback(async (): Promise<Result> => {
+        try {
+            const r = await mSQLite.isSecretStored();
+            if(r) {
+                return Promise.resolve(r);
+            } else {
+                return Promise.reject('Error in isSecretStored');
+            } 
+        } catch (err) {
+            return Promise.reject(err);
+        }
+
+    }, [mSQLite]);
+    const setEncryptionSecret = useCallback(async (passphrase: string): Promise<void> => {
+        if (passphrase == null || passphrase.length === 0) {
+            return Promise.reject(new Error('Must provide a passphrase'));
+        } 
+        try {
+            await mSQLite.setEncryptionSecret(passphrase);
+            return Promise.resolve();
+        } catch(err) {
+            return Promise.reject(err);
+        }
+
+    }, [mSQLite]);
+    const changeEncryptionSecret = useCallback(async (passphrase: string,
+        oldpassphrase: string): Promise<void> => {
+        if (passphrase == null || passphrase.length === 0) {
+            return Promise.reject(new Error('Must provide a passphrase'));
+        } 
+        if (oldpassphrase == null || oldpassphrase.length === 0) {
+            return Promise.reject(new Error('Must provide the old passphrase'));
+        } 
+        try {
+            await mSQLite.changeEncryptionSecret(passphrase, oldpassphrase);
+            return Promise.resolve();
+        } catch(err) {
+            return Promise.reject(err);
+        }
+
+    }, [mSQLite]);
 
     if (!availableFeaturesN.useSQLite) {
         return {
@@ -519,6 +583,9 @@ export const useSQLite = ({
             addSQLiteSuffix: featureNotAvailableError,
             deleteOldDatabases: featureNotAvailableError,
             checkConnectionsConsistency: featureNotAvailableError, 
+            isSecretStored: featureNotAvailableError,
+            setEncryptionSecret: featureNotAvailableError,
+            changeEncryptionSecret: featureNotAvailableError,             
             ...notAvailable
         };
     } else {
@@ -526,7 +593,9 @@ export const useSQLite = ({
             retrieveConnection, retrieveAllConnections, closeAllConnections,
             addUpgradeStatement, importFromJson, isJsonValid, copyFromAssets,
             isConnection, isDatabase, getDatabaseList, addSQLiteSuffix,
-            deleteOldDatabases, checkConnectionsConsistency, isAvailable: true};
+            deleteOldDatabases, checkConnectionsConsistency, 
+            isSecretStored, setEncryptionSecret, changeEncryptionSecret,
+            isAvailable: true};
     }
 
 }